+++ backend/kirikiri/src/main/java/co/kirikiri/service/dto/goalroom/response/GoalRoomForListResponse.java
@@ -1,4 +1,4 @@
-package co.kirikiri.service.dto.goalroom;
+package co.kirikiri.service.dto.goalroom.response;
 
 import co.kirikiri.service.dto.member.MemberResponse;
 import java.time.LocalDate;
