+++ backend/kirikiri/src/main/java/co/kirikiri/service/dto/roadmap/request/RoadmapSaveRequest.java
@@ -1,4 +1,4 @@
-package co.kirikiri.service.dto.roadmap;
+package co.kirikiri.service.dto.roadmap.request;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
