+++ backend/kirikiri/src/test/java/co/kirikiri/service/GoalRoomServiceTest.java
@@ -23,7 +23,7 @@ import co.kirikiri.persistence.goalroom.GoalRoomRepository;
 import co.kirikiri.service.dto.CustomPageRequest;
 import co.kirikiri.service.dto.PageResponse;
 import co.kirikiri.service.dto.goalroom.GoalRoomFilterTypeDto;
-import co.kirikiri.service.dto.goalroom.GoalRoomForListResponse;
+import co.kirikiri.service.dto.goalroom.response.GoalRoomForListResponse;
 import co.kirikiri.service.dto.member.MemberResponse;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
